test: UserWallet 관련 테스트 변경

+++ src/main/java/com/devkor/ifive/nadab/domain/wallet/core/repository/UserWalletRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface UserWalletRepository extends JpaRepository<UserWallet, Integer> {
+public interface UserWalletRepository extends JpaRepository<UserWallet, Long> {
 
     Optional<UserWallet> findByUserId(Long userId);
